Fixed to work with generic serial with size_ts

+++ embedded/tiva/transport_tiva_uart.c
@@ -18,18 +18,18 @@
 #endif
 
 
-static uint32_t uartPut(const uint8_t* data, uint32_t nData, void* usr)
+static size_t uartPut(const uint8_t* data, size_t nData, void* usr)
 {
     uint32_t uart_base = (uint32_t) usr;
-    uint32_t ret = 0;
+    size_t ret = 0;
     while (ret < nData && UARTCharPutNonBlocking(uart_base, data[ret])) ret++;
     return ret;
 }
 
-static uint32_t uartGet(uint8_t* data, uint32_t nData, void* usr)
+static size_t uartGet(uint8_t* data, size_t nData, void* usr)
 {
     uint32_t uart_base = (uint32_t) usr;
-    uint32_t ret = 0;
+    size_t ret = 0;
     while (ret < nData && UARTCharsAvail(uart_base))
         data[ret++] = UARTCharGet(uart_base);
     return ret;
